feat: Code refactoring and moving default key generation strategy (#11)
change db, separate the movie controller method, and change the key strategy in the id direct allocation method

+++ movierank/src/main/java/yangbrothers/movierank/job/RangePartitioner.java
@@ -12,20 +12,20 @@ public class RangePartitioner implements Partitioner {
     @Override
     public Map<String, ExecutionContext> partition(int gridSize) {
         HashMap<String, ExecutionContext> result = new HashMap<>();
-        int firstPage = 1;
-        int lastPage = 25;
+        long firstPage = 1;
+        long lastPage = 25;
         int count = 0;
 
         while (lastPage <= 100) {
-            int firstIndex = (firstPage - 1) * 100;
-            int lastIndex = firstIndex + 2501;
+            long firstIndex = (firstPage - 1) * 100;
+            long lastIndex = firstIndex + 2501;
             ExecutionContext executionContext = new ExecutionContext();
             result.put("partition " + count++, executionContext);
 
-            executionContext.putInt("firstPage", firstPage);
-            executionContext.putInt("lastPage", lastPage);
-            executionContext.putInt("firstIndex", firstIndex);
-            executionContext.putInt("lastIndex", lastIndex);
+            executionContext.putLong("firstPage", firstPage);
+            executionContext.putLong("lastPage", lastPage);
+            executionContext.putLong("firstIndex", firstIndex);
+            executionContext.putLong("lastIndex", lastIndex);
 
             firstPage += 25;
             lastPage += 25;
